==> notebooks/1 - Get Data.ipynb <==
# Starter Data Set

#### Imports

import pandas as pd

#### Intakes Data

intakes = pd.read_csv('https://data.austintexas.gov/api/views/wter-evkm/rows.csv?accessType=DOWNLOAD')

intakes.to_csv('../data/intakes.csv', index=False)

#### Outcomes Data

outcomes = pd.read_csv('https://data.austintexas.gov/api/views/9t4d-g238/rows.csv?accessType=DOWNLOAD')

outcomes.to_csv('../data/outcomes.csv', index=False)
