==> notebooks/2 - Intakes.ipynb <==
# Intakes Data

#### Imports

import pandas as pd
import janitor
import numpy as np
import re

#### Read in data

intakes = pd.read_csv('../data/intakes.csv')

#### Standardize column names

intakes = intakes.clean_names() #from janitor package

## Clean columns

#### Name

intakes['name'].str.startswith("*").value_counts(normalize=True)

#after doing some research, I couldn't any info on why some names have the *
intakes['name'] = intakes['name'].str.replace('*', '') 

#### Date time

intakes['datetime'] = pd.to_datetime(intakes['datetime'])

#### Year and Month

# we'll recreate this as a feature 
intakes = intakes.drop(columns='monthyear') 

intakes['intake_year'] = intakes['datetime'].dt.year

intakes['intake_month'] = intakes['datetime'].dt.month

#### Found location

intakes['found_location']

The pattern of this column is: <br>
    1) Street address that starts with numbers and ends with "in __ (TX)" <br>
    2) City that is formatted like "__ (TX)"

intakes['found_location_split'] = intakes['found_location'].str.split(' in ')

intakes['found_address'] = [x[0] if len(x) > 1 else np.nan for x in intakes['found_location_split']]

intakes['found_city'] = [x[1] if len(x) > 1 else x[0] for x in intakes['found_location_split']]

intakes['found_city'] = intakes['found_city'].str.replace(' \(TX\)', '')

# function returns the string if it starts with a number otherwise returns NaN

def contains_numbers(string):
    match = re.search('^\d', string)
    
    if match == None:
        return np.nan
    else:
        return string

intakes['found_address'] = [contains_numbers(str(x)) for x in intakes['found_address']]

#### Intake Type

intakes['intake_type'].value_counts()

#### Animal Type

intakes['animal_type'].value_counts()

#### Sex upon intake

intakes['sex_upon_intake'] = intakes['sex_upon_intake'].str.replace('Unknown', 'Unknown Unknown')

intakes['sex_split'] = intakes['sex_upon_intake'].str.split(' ')

intakes['spay_neuter'] = intakes['sex_upon_intake'].str.split(' ').str[0]

intakes['sex'] = intakes['sex_upon_intake'].str.split(' ').str[1]

#### Age upon intake

Strategy is to split the digit from the unit of time. Standardize to year. Make function to multiply digit based on unit of time

intakes['age_digit'] = [str.split(x, ' ')[0] for x in intakes['age_upon_intake']]

There are 7 negative values in the dataset that suggest the animal is less than __ years old (ex: -1 years old means less than 1 year old). Because these are estimates from the shelter, we'll convert them to positive numbers. Overall, the age numbers should not be taken as absolute truth

intakes['age_digit'] = intakes['age_digit'].str.replace('-', '')

intakes['age_digit'] = intakes['age_digit'].astype('int')

intakes['age_unit'] = [str.split(x, ' ')[1] for x in intakes['age_upon_intake']]

age_unit_dict = {
    'year':'years',
    'month':'months',
    'week': 'weeks',
    'day':'days'
}

intakes['age_unit'] = intakes['age_unit'].map(age_unit_dict).fillna(intakes['age_unit'])

def age_in_years(unit):
    if unit == 'years':
        return 1
    elif unit == 'months':
        return 12
    elif unit == 'weeks':
        return 52.143
    elif unit == 'days':
        return 365

intakes['age_conversion'] = [age_in_years(x) for x in intakes['age_unit']]



intakes['age_in_years'] = round(intakes['age_digit'] / intakes['age_conversion'], 2)

#### Categorical age grouping

def age_group(number):
    if number < 1:
        return 'under 1'
    elif 1 <= number < 2.99:
        return 'between 1 and 3'
    elif 3 <= number < 4.99:
        return 'between 3 and 5'
    else: 
        return 'over 5'

intakes['age_group'] = intakes['age_in_years'].apply(age_group)

#### Breed

#Black and Tan Hounds' is listed as Black/Tan Hound and that naming convention causes flag and/or split issues

intakes['breed'] = intakes['breed'].str.replace('Black/Tan Hound', 'Black and Tan Hound')

#Mixed breeds either contain "Mix" or "/"
intakes['mix_flag'] = np.where((intakes['breed'].str.contains(' Mix')) | (intakes['breed'].str.contains('\/')), True, False)

# Flag for dog breed containing Pit Bull
intakes['pit_flag'] = np.where(intakes['breed'].str.contains('Pit Bull'), True, False)

intakes['breed_split'] = intakes['breed'].str.split('\/')

#sorts the list of mixed breeds alphabetically to avoid combinations from ordering of breed

intakes['breed_split_sort'] = [sorted(x) if len(x) > 1 else x for x in intakes['breed_split']]

intakes['breed_1'] = intakes['breed_split_sort'].str[0]
intakes['breed_2'] = intakes['breed_split_sort'].str[1]

#remove mix

intakes['breed_1'] = intakes['breed_1'].str.replace(' Mix', '')
intakes['breed_2'] = intakes['breed_2'].str.replace(' Mix', '')

#### Color

intakes['color_1'] = intakes['color'].str.split('\/').str[0]

intakes['color_2'] = intakes['color'].str.split('\/').str[1]

## Additional Feature Creation

#### Category
This will be used for an intermediate table along the way to properly align intakes and outcomes

intakes['category'] = 'intake'

#### Dropping unnecessary columns

cols_to_drop = [
    'found_location',
    'sex_upon_intake',
    'age_upon_intake',
    'found_location_split',
    'sex_split',
    'age_digit',
    'age_unit',
    'age_conversion',
    'breed_split',
    'breed_split_sort'
]

intakes = intakes.drop(columns=cols_to_drop)

intakes.to_csv('../data/intakes_clean.csv', index=False)

intakes['datetime'].sort_values()
