==> notebooks/3 - Outcomes.ipynb <==
## Outcomes

#### Imports

import pandas as pd
import janitor
import numpy as np

#### Read in Data

outcomes = pd.read_csv('../data/outcomes.csv', parse_dates=['DateTime'])

outcomes = outcomes.clean_names()

#dropping information available in the intake data

columns_to_drop = [
    'name',
    'monthyear',
    'date_of_birth',
    'animal_type',
    'breed',
    'color'
]

outcomes = outcomes.drop(columns=columns_to_drop)

outcomes.head()

#### Outcome Type

outcomes['outcome_type'].value_counts(normalize=True)

#### Outcomes Subtype

outcomes['outcome_subtype'].value_counts()

#### Sex Upon Outcome

#pattern is neuter/spay and gender, so I double the unknowns to make the split easier
outcomes['sex_upon_outcome'] = outcomes['sex_upon_outcome'].str.replace('Unknown', 'Unknown Unknown')

outcomes['sex_split'] = outcomes['sex_upon_outcome'].str.split(' ')

outcomes['spay_neuter'] = outcomes['sex_split'].str[0]
outcomes['sex'] = outcomes['sex_split'].str[1]

#### Age Upon Outcome

outcomes['age_digit'] = outcomes['age_upon_outcome'].str.split(' ').str[0]
outcomes['age_unit'] = outcomes['age_upon_outcome'].str.split(' ').str[1]

outcomes['age_digit'] = outcomes['age_digit'].astype('float')

age_unit_dict = {
    'year':'years',
    'month':'months',
    'week': 'weeks',
    'day':'days'
}

outcomes['age_unit'] = outcomes['age_unit'].map(age_unit_dict).fillna(outcomes['age_unit'])

def age_in_years(unit):
    if unit == 'years':
        return 1
    elif unit == 'months':
        return 12
    elif unit == 'weeks':
        return 52.143
    elif unit == 'days':
        return 365

outcomes['age_conversion'] = [age_in_years(x) for x in outcomes['age_unit']]

outcomes['age_in_years'] = round((outcomes['age_digit'] / outcomes['age_conversion']), 2)

## Feature Creation

#### Outcomes Count

outcomes['category'] = 'outcome'

#### Drop Columns

cols_to_drop = [
    'sex_upon_outcome',
    'sex_split',
    'age_upon_outcome',
    'age_digit',
    'age_unit',
    'age_conversion'
]

outcomes = outcomes.drop(columns=cols_to_drop)

## Data to CSV

outcomes.to_csv('../data/outcomes_cleaned.csv', index=False)
