==> notebooks/4 - Intakes and Outcomes Merge.ipynb <==
# Intakes and Outcomes Merge

#### Imports

import pandas as pd

#### Read in Data

intakes = pd.read_csv('../data/intakes_clean.csv', parse_dates=['datetime'])

outcomes = pd.read_csv('../data/outcomes_cleaned.csv', parse_dates=['datetime'])

#### Merge Datasets

intake_merge = intakes[['animal_id', 'datetime', 'category']]

outcome_merge = outcomes[['animal_id', 'datetime', 'category']]

intake_merge.head()

outcome_merge.head()

intermediate = pd.concat([intake_merge, outcome_merge]).sort_values('datetime')

#this will allow us to get intakes and outcomes matched and on the same row

intermediate['time_shift'] = intermediate.groupby('animal_id').datetime.shift(-1)
intermediate['next_event'] = intermediate.groupby('animal_id').category.shift(-1)

intermediate.head()

#ensures we only get intake to outcome results
intermediate = intermediate[(intermediate['category'] == 'intake') & (intermediate['next_event'] == 'outcome')]

intakes_merge = intermediate.merge(intakes, on=['animal_id', 'datetime'], suffixes=['_inter', '_intakes'])

intakes_outcomes = intakes_merge.merge(outcomes, 
                        left_on=['animal_id', 'time_shift'], 
                        right_on=['animal_id', 'datetime'],
                        suffixes=['_inter', '_outcomes'])

#### Reordering, Dropping, and Renaming Columns

columns_to_keep = [
    'animal_id',
    'datetime_inter',
    'name',
    'intake_type',
    'intake_condition',
    'animal_type',
    'sex_inter',
    'spay_neuter_inter',
    'age_in_years_inter',
    'age_group',
    'breed',
    'breed_1',
    'breed_2',
    'mix_flag',
    'pit_flag',
    'color',
    'color_1',
    'color_2',
    'found_address',
    'found_city',
    'datetime_outcomes',
    'outcome_type',
    'outcome_subtype',
    'sex_outcomes',
    'spay_neuter_outcomes',
    'age_in_years_outcomes'
]

intakes_outcomes = intakes_outcomes[columns_to_keep]

intakes_outcomes = intakes_outcomes.rename(
                                            columns={
                                                    'datetime_inter':'intake_datetime', 
                                                    'sex_inter':'intake_sex',
                                                    'spay_neuter_inter':'intake_spay_neuter',
                                                    'age_in_years_inter':'intake_age_in_years',
                                                    'sex_outcomes':'outcome_sex',
                                                    'spay_neuter_outcomes':'outcome_spay_neuter',
                                                    'age_in_years_outcomes':'outcome_age_in_years'
                                                    }
                                            )

#### Filter Dataset

dogs = intakes_outcomes[intakes_outcomes['animal_type'] == 'Dog']

#### Save csv

dogs.to_csv('../data/dogs.csv', index=False)
