==> notebooks/8 - Even Simpler Model.ipynb <==
#### Imports

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report

#### Read in Dataset

dogs = pd.read_csv('../data/dogs.csv', parse_dates=['intake_datetime'])

#removing low frequency events

dogs = dogs[~dogs['outcome_type'].isin(['Died', 'Disposal', 'Missing', 'Rto-Adopt'])]

#### Define Features

Trying an even simpler model based on intake_type and intake_age_in_years

features = [
    'intake_type',
    'intake_age_in_years',
]

target = ['outcome_type']

#14 outcomes are NaN so we drop those

dogs = dogs.dropna(subset=['outcome_type'])

X = dogs[features]
y = dogs['outcome_type']

X = pd.get_dummies(data=X, 
                   columns=['intake_type'],
                  drop_first=True)

#### Train, Test, Split

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=63)

#### Fit RandomForest Model

clf = RandomForestClassifier()

clf.fit(X_train, y_train)

#### Results

clf.predict_proba(X_test)

accuracy_score(y_test, clf.predict(X_test))

print(classification_report(y_test, clf.predict(X_test)))
