==> notebooks/9 - CatBoost Model.ipynb <==
#### Imports

import pandas as pd
import numpy as np
from pandas.api.types import is_numeric_dtype
from catboost import CatBoostClassifier, Pool 
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report

#### Read in Dataset

dogs = pd.read_csv('../data/dogs.csv')

#### Feature Selection

dogs = dogs.dropna(subset=['outcome_type', 'intake_sex'])

features = [
    'intake_type',
    'intake_condition',
    'intake_sex',
    'intake_spay_neuter',
    'intake_age_in_years',
    'breed',
    'color',
    'found_city',
]

target = ['outcome_type']

X = dogs[features]
y = dogs[target]

#CatBoost requires the specification of categorical columns. This function gets the list of categorical indicies

def get_categorical_indicies(X):
    cats = []
    for col in X.columns:
        if is_numeric_dtype(X[col]):
            pass
        else:
            cats.append(col)
    cat_indicies = []
    for col in cats:
        cat_indicies.append(X.columns.get_loc(col))
    return cat_indicies

categorical_indicies = get_categorical_indicies(X)

categorical_indicies

#CatBoost needs the categorical columns to be converted to the categorical datatype

def convert_cats(X):
    cats = []
    for col in X.columns:
        if is_numeric_dtype(X[col]):
            pass
        else:
            cats.append(col)
    cat_indicies = []
    for col in cats:
        X[col] = X[col].astype('category')

convert_cats(X)

#### Train, Test, Split and Pooling

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=63)

train_dataset = Pool(X_train, 
                     y_train,
                     cat_features=categorical_indicies
                    )
test_dataset = Pool(X_test, 
                    y_test,
                    cat_features=categorical_indicies
                   )

#### Model Training and Prediction

clf = CatBoostClassifier()

clf.fit(train_dataset)

preds = clf.predict(test_dataset)

print(preds)

#### Model Performance

print(classification_report(y_test, preds))

clf.get_feature_importance(prettified=True)
