==> notebooks/15 - Unseen Data.ipynb <==
import pandas as pd
import numpy as np
import janitor

### Unseen Intakes

unseen_intakes = pd.read_csv('https://data.austintexas.gov/api/views/wter-evkm/rows.csv?accessType=DOWNLOAD')

unseen_intakes = unseen_intakes.clean_names()

#### Subset data to dogs and only get new data

unseen_intakes = unseen_intakes[unseen_intakes['animal_type'] == 'Dog']

unseen_intakes['datetime'] = pd.to_datetime(unseen_intakes['datetime'])

unseen_intakes = unseen_intakes[unseen_intakes['datetime'].dt.date > pd.to_datetime('2021-05-16')]

#### Add year and month columns

unseen_intakes['intake_year'] = unseen_intakes['datetime'].dt.year
unseen_intakes['intake_month'] = unseen_intakes['datetime'].dt.month

#### Get sex and spay/neuter split

unseen_intakes['sex_upon_intake'] = unseen_intakes['sex_upon_intake'].str.replace('Unknown', 'Unknown Unknown')
unseen_intakes['sex_split'] = unseen_intakes['sex_upon_intake'].str.split(' ')
unseen_intakes['spay_neuter'] = unseen_intakes['sex_upon_intake'].str.split(' ').str[0]
unseen_intakes['sex'] = unseen_intakes['sex_upon_intake'].str.split(' ').str[1]

#### Age in Years

#age digit
unseen_intakes['age_digit'] = [str.split(x, ' ')[0] for x in unseen_intakes['age_upon_intake']]
unseen_intakes['age_digit'] = unseen_intakes['age_digit'].str.replace('-', '')
unseen_intakes['age_digit'] = unseen_intakes['age_digit'].astype('int')

#age unit
unseen_intakes['age_unit'] = [str.split(x, ' ')[1] for x in unseen_intakes['age_upon_intake']]
age_unit_dict = {
    'year':'years',
    'month':'months',
    'week': 'weeks',
    'day':'days'
}

unseen_intakes['age_unit'] = unseen_intakes['age_unit'].map(age_unit_dict).fillna(unseen_intakes['age_unit'])

def age_in_years(unit):
    if unit == 'years':
        return 1
    elif unit == 'months':
        return 12
    elif unit == 'weeks':
        return 52.143
    elif unit == 'days':
        return 365

unseen_intakes['age_conversion'] = [age_in_years(x) for x in unseen_intakes['age_unit']]
unseen_intakes['age_in_years'] = round(unseen_intakes['age_digit'] / unseen_intakes['age_conversion'], 2)

#### Breed

unseen_intakes['breed'] = unseen_intakes['breed'].str.replace('Black/Tan Hound', 'Black and Tan Hound')

#Mixed breeds either contain "Mix" or "/"
unseen_intakes['mix_flag'] = np.where((unseen_intakes['breed'].str.contains(' Mix')) | (unseen_intakes['breed'].str.contains('\/')), True, False)

#### Category

unseen_intakes['category'] = 'intake'

columns_to_keep = [
    'animal_id',
    'intake_year',
    'intake_month',
    'datetime',
    'intake_type',
    'intake_condition',
    'spay_neuter',
    'age_in_years',
    'breed',
    'category'
]

unseen_intakes = unseen_intakes[columns_to_keep]

### Unseen Outcomes

unseen_outcomes = pd.read_csv('https://data.austintexas.gov/api/views/9t4d-g238/rows.csv?accessType=DOWNLOAD')

unseen_outcomes = unseen_outcomes.clean_names()

#### Subset data

unseen_outcomes = unseen_outcomes[unseen_outcomes['animal_type'] == 'Dog']

unseen_outcomes['datetime'] = pd.to_datetime(unseen_outcomes['datetime'])

unseen_outcomes = unseen_outcomes[unseen_outcomes['datetime'].dt.date > pd.to_datetime('2021-05-16')]

#### Category

unseen_outcomes['category'] = 'outcome'

#### Selecting columns

unseen_outcomes = unseen_outcomes[['animal_id', 'datetime', 'outcome_type', 'category']]

### Merge

i_merge = unseen_intakes[['animal_id', 'datetime', 'category']]
o_merge = unseen_outcomes[['animal_id', 'datetime', 'category']]

intermediate = pd.concat([i_merge, o_merge]).sort_values('datetime')

intermediate['time_shift'] = intermediate.groupby('animal_id').datetime.shift(-1)
intermediate['next_event'] = intermediate.groupby('animal_id').category.shift(-1)

intermediate

intermediate = intermediate[(intermediate['category'] == 'intake') & (intermediate['next_event'] == 'outcome')]

intermediate

intermediate = intermediate.merge(unseen_intakes, on=['animal_id', 'datetime'], suffixes=['_inter', '_intakes'])

unseen_merged = intermediate.merge(unseen_outcomes, left_on=['animal_id', 'time_shift'], right_on=['animal_id', 'datetime'])

#### Subset to "Normal" intake condition

unseen_merged = unseen_merged[unseen_merged['intake_condition'] == 'Normal']

#### Rename Columns

unseen_merged = unseen_merged.rename(columns={'spay_neuter':'intake_spay_neuter', 
                                              'breed_x':'breed',
                                             'age_in_years':'intake_age_in_years'})

unseen_merged.columns

#### Select Columns

unseen_merged = unseen_merged[['intake_type', 'intake_year', 'intake_month', 'intake_spay_neuter', 'breed', 'intake_age_in_years', 'outcome_type']]

unseen_merged.to_csv('../data/unseen_data', index=False)
